refactor(foundation,redis): relocate RedisCodeRepository [Theme C/E]
Phase 10 Q4: move RedisCodeRepository + its test from
@o3co/auth-provider-foundation to @o3co/auth-provider-redis. Foundation
reduces to HTTP-only adapters; its `redis` peer dependency is removed.

- packages/redis/src/code-repository.mts: class + builder
- packages/foundation/src/index.mts: registerBuiltinAdapters keeps the
  v0.4.x signature (codeFactory now optional/no-op) for backward compat
- packages/foundation/package.json: drop `redis` peer dep + meta
- packages/redis/package.json: add `redis` v5 as optional peer + dev

Migration: consumers wire the redis CodeRepository via:
  import { redisCodeRepositoryBuilder } from "@o3co/auth-provider-redis";
  codeFactory.register("redis", redisCodeRepositoryBuilder);

Module-pattern wrapper for the `codeRepository` ComponentMap slot is
intentionally deferred to a separate "legacy-slot module-parity" PR.
