+++ packages/foundation/package.json
@@ -35,19 +35,12 @@
 		}
 	},
 	"peerDependencies": {
-		"@o3co/auth-provider-core": "^0.0.0",
-		"redis": "^5.10.0"
-	},
-	"peerDependenciesMeta": {
-		"redis": {
-			"optional": true
-		}
+		"@o3co/auth-provider-core": "^0.0.0"
 	},
 	"devDependencies": {
 		"@o3co/auth-provider-core": "workspace:*",
 		"@types/node": "^25.1.0",
 		"msw": "^2.12.14",
-		"redis": "^5.10.0",
 		"pino": "^10.3.1",
 		"vitest": "^4.1.2",
 		"@vitest/coverage-v8": "^4.1.2"
